+++ astyanax-test/src/main/java/com/netflix/astyanax/cql/test/PreparedStatementTests.java
@@ -299,7 +299,7 @@ public class PreparedStatementTests extends ReadTests {
         	String rowKey = Character.toString(keyName);
         	ColumnListMutation<String> colMutation = m.withRow(CF_ROW_RANGE, rowKey);
               for (char cName = 'a'; cName <= 'z'; cName++) {
-            	  colMutation.putColumn(Character.toString(cName), (int) (cName - 'a') + 1, null);
+            	  colMutation.putColumn(Character.toString(cName), cName - 'a' + 1, null);
               }
               m.withCaching(true);
               m.execute();
